pyproject.toml
[tool.poetry]
name = "wb"
version = "0.1.0"
description = ""
authors = ["Divider"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.10.0"
python-dotenv = "^1.0.1"
aioschedule = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
